+++ src/smart-components/Channel/components/MessageInput/VoiceMessageInputWrapper.tsx
@@ -4,7 +4,7 @@ import './voice-message-wrapper.scss';
 
 import useSendbirdStateContext from '../../../../hooks/useSendbirdStateContext';
 import { useLocalization } from '../../../../lib/LocalizationContext';
-import { useVoicePlayer } from '../../../../hooks/useVoicePlayer/useVoicePlayer';
+import { useVoicePlayer } from '../../../../hooks/VoicePlayer/useVoicePlayer';
 import { useVoiceRecorder } from '../../../../hooks/VoiceRecorder/useVoiceRecorder';
 import { isDisabledBecauseFrozen, isDisabledBecauseMuted } from '../../context/utils';
